[Fix] Many bugs [Add] ESS filewriter flatbuffer control
- All Instance parameters are now treated as Expr expression objects as
  late as possible in the construction of the equivalent NX* class
  object. This potentially enables simplifying expressions to constant
  values or simpler expressions before writing the constants or
expressions to the NeXus structure.
- An attempt is made to parse the C declarations and initialization of
  the instrument to allow extraction of parameter values for evaluation
of the component instance parameter expressions. The C expression parser
is not fully featured, however, so this will likely fail.
- If a NX* parameter ultimately depends on an *Instrument* parameter
  (one set from the command line) it will be replaced by an NXcollection
object with members 'expression' and ESS filewriter module dicts for
links to the parameter NXlog entries -- this requires that the NXlog
module dicts are added (probably later) as well, and that a McStas to
EPICS forwarder component is included in the instrument to enable
sending the runtime instrument parameter values on to Kafka.